backend/requirements.txt
black
flake8
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-dotenv==1.0.0
pydantic==1.10.15
SQLAlchemy==2.0.29
sqlmodel==0.0.16
web3==6.13.0
eth-account==0.10.0
itsdangerous==2.2.0
python-multipart==0.0.9
redis
pyjwt
cryptography
